requirements.txt
numpy
google-cloud-bigquery
opencv-python
queue; python_version < "3.0"
threading; python_version < "3.0"
torch
transformers
Pillow